fix(): icon sizes
